Code Refactoring 1. Decompose Tile.java. 2. Change tile project structure.
