+++ src/com.chess.engine.board/board/tiles/OccupiedTile.java
@@ -1,0 +1,23 @@
+package com.chess.engine.board.board.tiles;
+
+import com.chess.engine.board.pieces.Piece;
+
+public final class OccupiedTile extends Tile {
+
+    private final Piece pieceOnTile;
+
+    public OccupiedTile(int TileCoordinate, Piece pieceOnTile) {
+        super(TileCoordinate);
+        this.pieceOnTile = pieceOnTile;
+    }
+
+    @Override
+    public boolean isTileOccupied() {
+        return true;
+    }
+
+    @Override
+    public Piece getPiece() {
+        return this.pieceOnTile;
+    }
+}
